+++ types/html-to-image/types.d.ts
@@ -68,7 +68,7 @@ export interface Options {
     /**
      * The preferred font format. If specified all other font formats are ignored.
      */
-    preferredFontFormat?: 'woff' | 'woff2' | 'truetype' | 'opentype' | 'embedded-opentype' | 'svg' | string;
+    preferredFontFormat?: string;
     /**
      * A CSS string to specify for font embeds. If specified only this CSS will
      * be present in the resulting image. Use with `getFontEmbedCSS()` to